Changed to follow new module system call design
