+++ src/sys_window_struct.h
@@ -1,0 +1,16 @@
+/*
+Copyright (C) 2016 The University of Notre Dame
+This software is distributed under the GNU General Public License.
+See the file LICENSE for details.
+*/
+
+#ifndef SYS_WINDOW_STRUCT_H
+#define SYS_WINDOW_STRUCT_H
+
+struct graphics_color {
+    uint8_t r;
+    uint8_t g;
+    uint8_t b;
+};
+
+#endif
